7.2.6: Improve DRS labelling
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/browser/trunk@89584 735d13b6-9817-0410-8766-e36946ffe9aa

+++ lib/class.tx_browser_befilter_ts.php
@@ -160,7 +160,7 @@ class tx_browser_befilter_ts
     } // LOOP each marker value in the page TSconfig
 
     $marker = '###CURRENT_PID###'; // #i0189, 150822, dwildt, 3+
-    $value = t3lib_div::_GP( 'id' );
+    $value = ( int ) t3lib_div::_GP( 'id' );
     $conf[ 'foreign_table_where' ] = str_replace( $marker, $value, $conf[ 'foreign_table_where' ] );
 
     return $conf;
